Cache 2

See merge request CCP4/jsCoFE!174

+++ js-server/server.utils.js
@@ -2,7 +2,7 @@
 /*
  *  =================================================================
  *
- *    23.10.24   <--  Date of Last Modification.
+ *    02.11.24   <--  Date of Last Modification.
  *                   ~~~~~~~~~~~~~~~~~~~~~~~~~~~~
  *  -----------------------------------------------------------------
  *
@@ -13,6 +13,7 @@
  *  **** Content :  Server-side utility functions
  *       ~~~~~~~~~
  *
+ *        function configureCache   ( ncache )
  *        function fileExists       ( fpath )
  *        function isSymbolicLink   ( fpath )
  *        function dirExists        ( fpath )
@@ -73,6 +74,7 @@ const child_process = require('child_process');
 
 const class_map     = require('./server.class_map');
 const task_t        = require('../js-common/tasks/common.tasks.template');
+const cache         = require('./server.cache');
 // const com_utils     = require('../js-common/common.utils');
 
 //  prepare log
@@ -80,9 +82,28 @@ const log = require('./server.log').newLog(14);
 
 const _is_windows = /^win/.test(process.platform);
 
+
 // ==========================================================================
 
+var cache_enabled = false;
+
+function configureCache ( ncache )  {
+// ncache is estimated number of users working simultaneously in the system
+  cache.configureCache ( ncache );
+  cache_enabled = cache.isCacheEnabled();
+}
+
 function fileExists ( fpath )  {
+  try {
+    if (cache_enabled && (cache.itemExists(fpath)>0))
+      return true;
+    return fs.lstatSync(fpath); // || fs.lstatSync(path);
+  } catch (e)  {
+    return null;
+  }
+}
+
+function fileStat ( fpath )  {
   try {
     return fs.lstatSync(fpath); // || fs.lstatSync(path);
   } catch (e)  {
@@ -124,6 +145,8 @@ function fileSize ( fpath ) {
 
 function removeFile ( fpath ) {
   try {
+    if (cache_enabled) 
+      cache.removeItem ( fpath );
     fs.unlinkSync ( fpath );
     return true;
   } catch (e)  {
@@ -134,6 +157,14 @@ function removeFile ( fpath ) {
 
 function readString ( fpath )  {
   try {
+    if (cache_enabled)  {
+      let json_str = cache.getItem ( fpath );
+      if (!json_str)  {
+        json_str = fs.readFileSync(fpath).toString();
+        cache.putItem ( fpath,json_str );
+      }
+      return json_str;
+    }
     return fs.readFileSync(fpath).toString();
   } catch (e)  {
     return null;
@@ -147,7 +178,7 @@ function makeSymLink ( pathToTarget,pathToOrigin )  {
           fs.symlinkSync ( pathToOrigin,pathToTarget,'junction' );
     else  fs.symlinkSync ( pathToOrigin,pathToTarget );
   } catch (e)  {
-    return null;
+    return false;
   }
   return true;
 }
@@ -155,10 +186,20 @@ function makeSymLink ( pathToTarget,pathToOrigin )  {
 
 function readObject ( fpath )  {
   try {
+    if (cache_enabled)  {
+      let json_str = cache.getItem ( fpath );
+      if (!json_str)  {
+        json_str = fs.readFileSync(fpath).toString();
+        cache.putItem ( fpath,json_str );
+      } 
+      return JSON.parse ( json_str );
+    }
     return JSON.parse ( fs.readFileSync(fpath).toString() );
   } catch (e)  {
-    if (e.code !== 'ENOENT')
+    if (e.code !== 'ENOENT')  {
       log.error ( 10, e.message + ' when loading ' + fpath );
+      console.error ( e );
+    }
     return null;
   }
 }
@@ -166,6 +207,14 @@ function readObject ( fpath )  {
 
 function readClass ( fpath ) {  // same as object but with class functions
   try {
+    if (cache_enabled)  {
+      let json_str = cache.getItem ( fpath );
+      if (!json_str)  {
+        json_str = fs.readFileSync(fpath).toString();
+        cache.putItem ( fpath,json_str );
+      }
+      return class_map.getClassInstance ( json_str );
+    }
     return class_map.getClassInstance ( fs.readFileSync(fpath).toString() );
   } catch (e)  {
     return null;
@@ -175,10 +224,20 @@ function readClass ( fpath ) {  // same as object but with class functions
 
 function writeString ( fpath,data_string )  {
   try {
-    fs.writeFileSync ( fpath,data_string );
+    if (cache_enabled && cache.putItem(fpath,data_string))  {
+      // was put into cache, use asynchronous write
+      fs.writeFile ( fpath,data_string,function(err){
+        if (err)  {
+          log.error ( 20,'cannot write file ' + fpath );
+          console.error(err);
+        }
+      });
+    } else  {
+      fs.writeFileSync ( fpath,data_string );
+    }
     return true;
   } catch (e)  {
-    log.error ( 20,'cannot write file ' + fpath +
+    log.error ( 21,'cannot write file ' + fpath +
                    ' error: ' + JSON.stringify(e) );
     console.error(e);
     return false;
@@ -187,6 +246,8 @@ function writeString ( fpath,data_string )  {
 
 
 function appendString ( fpath,data_string )  {
+  if (cache_enabled) 
+    cache.removeItem ( fpath );
   try {
     fs.appendFileSync ( fpath,data_string );
     return true;
@@ -199,7 +260,10 @@ function appendString ( fpath,data_string )  {
 }
 
 
-function writeObject ( fpath,dataObject )  {
+function writeObject ( fpath,dataObject,force_sync=false,callback_func=null )  {
+
+// if (fpath.endsWith('projects.list'))
+//   console.log ( ' >>>>> write project list' );
 
   let json_str = '';
   try {
@@ -213,10 +277,25 @@ function writeObject ( fpath,dataObject )  {
   }
 
   try {
-    fs.writeFileSync ( fpath,json_str );
+    if (((!force_sync) && cache_enabled && cache.putItem(fpath,json_str)) ||
+        callback_func)  {
+      // was put into cache, use asynchronous write
+      fs.writeFile ( fpath,json_str,function(err){
+    // if (fpath.endsWith('projects.list'))
+    //   console.log ( ' >>>>>\n' + json_str );
+        if (err)  {
+          log.error ( 41,'cannot write file ' + fpath );
+          console.error(err);
+        }
+        if (callback_func)
+          callback_func ( err );
+      });
+    } else  {
+      fs.writeFileSync ( fpath,json_str );
+    }
     return true;
   } catch (e)  {
-    log.error ( 41,'cannot write file ' + fpath );
+    log.error ( 42,'cannot write file ' + fpath );
     console.error(e);
     return false;
   }
@@ -247,6 +326,10 @@ function moveFile ( old_path,new_path )  {
   // this function should be used in asynchronous code; use in synchronous code
   // must be limited only when source and destination are known to be in
   // the same partition
+
+  if (cache_enabled) 
+    cache.removeItem ( old_path );
+
   try {
     if (_is_windows && fileExists(new_path))
       fs.unlinkSync ( new_path );
@@ -255,6 +338,7 @@ function moveFile ( old_path,new_path )  {
     log.error ( 60,'error: ' + JSON.stringify(e) );
     console.error(e);
   }
+
   try {
     fs.moveSync ( old_path,new_path,{'overwrite':true} );
 //    fs.renameSync ( old_path,new_path );
@@ -269,6 +353,7 @@ function moveFile ( old_path,new_path )  {
     log.error ( 61,'error: ' + JSON.stringify(e) );
     return false;
   }
+
 }
 
 
@@ -347,6 +432,7 @@ function mkDir ( dirPath )  {
   }
 }
 
+
 function mkDir_check ( dirPath )  {
 // attempts to create directory and returns:
 //     0 : if directory was created 
@@ -396,9 +482,17 @@ function mkPath ( dirPath )  {
 }
 
 
+function flushDirCache ( dir_path )  {
+  if (cache_enabled) 
+      cache.removeDirItems ( dir_path );
+}
+
 function removePathAsync ( dir_path,tmp_dir='' )  {
-let rc   = true;
-let stat = fileExists(dir_path);
+let rc = true;
+
+  flushDirCache ( dir_path );
+  
+  let stat = fileStat ( dir_path );
 
   if (stat)  {
     try {
@@ -442,9 +536,11 @@ let stat = fileExists(dir_path);
 
 
 function removePath ( dir_path )  {
-let rc   = true;
-let stat = fileExists(dir_path);
+let rc = true;
 
+  flushDirCache ( dir_path );
+
+  let stat = fileStat ( dir_path );
   if (stat && stat.isSymbolicLink())  {
     fs.unlinkSync ( dir_path );
   } else if (stat)  {
@@ -457,31 +553,6 @@ let stat = fileExists(dir_path);
       rc = false;
     }
 
-  //   fs.readdirSync(dir_path).forEach(function(file,index){
-  //     let curPath = path.join ( dir_path,file );
-  //     let curstat = fileExists ( curPath );
-  //     if (!curstat)  {
-  //       log.error ( 82,'cannot stat path ' + curPath );
-  //       rc = false;
-  //     } else if (curstat.isDirectory()) { // recurse
-  //       removePath ( curPath );
-  //     } else { // delete file
-  //       try {
-  //         fs.unlinkSync ( curPath );
-  //       } catch (e)  {
-  //         log.error ( 83,'cannot remove file ' + curPath +
-  //                        ' error: ' + JSON.stringify(e) );
-  //         rc = false;
-  //       }
-  //     }
-  //   });
-  //   try {
-  //     fs.rmdirSync ( dir_path );
-  //   } catch (e)  {
-  //     log.error ( 9,'cannot remove directory ' + dir_path +
-  //                   ' error: ' + JSON.stringify(e) );
-  //     rc = false;
-  //   }
   }
 
   return rc;  // false if there were errors
@@ -492,6 +563,9 @@ let stat = fileExists(dir_path);
 function moveDir ( old_path,new_path,overwrite_bool )  {
   // uses sync mode, which is Ok for source/destinations being on the same
   // file systems; use not-synced version when moving across devices
+
+  flushDirCache ( old_path );
+
   try {
     if (_is_windows && overwrite_bool && fileExists(new_path))
       removePathAsync ( new_path );
@@ -517,7 +591,11 @@ function moveDir ( old_path,new_path,overwrite_bool )  {
   }
 }
 
+
 function moveDirAsync ( old_path,new_path,overwrite_bool,callback_func )  {
+
+  flushDirCache ( old_path );
+
   try {
     if (_is_windows && overwrite_bool && fileExists(new_path))
       removePathAsync ( new_path );
@@ -526,6 +604,7 @@ function moveDirAsync ( old_path,new_path,overwrite_bool,callback_func )  {
     log.error ( 130,'error: ' + JSON.stringify(e) );
     console.error(e);
   }
+
   fs.move ( old_path,new_path,{'overwrite':overwrite_bool},function(err){
     if (err)  {
       let old_exist = '(non-existing)';
@@ -539,16 +618,21 @@ function moveDirAsync ( old_path,new_path,overwrite_bool,callback_func )  {
     }
     callback_func(err);
   });
+
 }
+
 
 function cleanDir ( dir_path,exclude=[] ) {
   // removes everything in the directory, but does not remove it
+
+  flushDirCache ( dir_path );
+
   let rc = true;
   if (fileExists(dir_path))  {
     fs.readdirSync(dir_path).forEach(function(file,index){
       if (!exclude.includes(file))  {
         let curPath = path.join ( dir_path,file );
-        let curstat = fileExists ( curPath );
+        let curstat = fileStat  ( curPath );
         if (!curstat)  {
           log.error ( 140,'cannot stat path ' + curPath );
           rc = false;
@@ -574,10 +658,13 @@ function cleanDirExt ( dir_path,fext )  {
   // removes all files with given extension recursively in the directory,
   // but does not remove any directories, even if they are empty
   let rc = true;
+
+  flushDirCache ( dir_path );
+
   if (fileExists(dir_path))  {
     fs.readdirSync(dir_path).forEach(function(file,index){
       let curPath = path.join ( dir_path,file );
-      let curstat = fileExists ( curPath );
+      let curstat = fileStat  ( curPath );
       if (!curstat)  {
         log.error ( 150,'cannot stat path ' + curPath );
         rc = false;
@@ -594,17 +681,23 @@ function cleanDirExt ( dir_path,fext )  {
       }
     });
   }
+
   return rc;  // false if there were errors
+
 }
 
 
 function removeSymLinks ( dir_path )  {
 // removes all symbolic links recursively in the directory
   let rc = true;
+
+  if (cache_enabled) 
+    cache.removeItem ( dir_path );
+
   if (fileExists(dir_path))  {
     fs.readdirSync(dir_path).forEach(function(file,index){
       let curPath = path.join ( dir_path,file );
-      let curstat = fileExists ( curPath );
+      let curstat = fileStat  ( curPath );
       if (!curstat)  {
         log.error ( 160,'cannot stat path ' + curPath );
         rc = false;
@@ -614,6 +707,8 @@ function removeSymLinks ( dir_path )  {
         try {
           let fpath = fs.readlinkSync ( curPath );
           fs.unlinkSync ( curPath );
+          if (cache_enabled) 
+            cache.removePathItem ( curPath );
           if (fs.existsSync(fpath))
             fs.copyFileSync ( fpath, curPath );
         } catch (e)  {
@@ -623,7 +718,9 @@ function removeSymLinks ( dir_path )  {
         }
     });
   }
+
   return rc;  // false if there were errors
+
 }
 
 
@@ -633,7 +730,7 @@ function getDirectorySize ( dir_path )  {
     if (fileExists(dir_path))  {
       fs.readdirSync(dir_path).forEach(function(file,index){
         let curPath = path.join ( dir_path,file );
-        let curstat = fileExists ( curPath );
+        let curstat = fileStat ( curPath );
         if (curstat)  {
           if (curstat.isDirectory())  { // recurse
             size += getDirectorySize ( curPath );
@@ -662,7 +759,7 @@ function searchTree ( dir_path,filename,matchKey ) {
     if (fileExists(dir_path))  {
       fs.readdirSync(dir_path).forEach(function(file,index){
         let curPath = path.join ( dir_path,file );
-        let curstat = fileExists ( curPath );
+        let curstat = fileStat  ( curPath );
         if (curstat && curstat.isDirectory()) { // recurse
           filepaths = filepaths.concat ( searchTree(curPath,filename,matchKey) );
         } else if (((matchKey==0) && (file==filename)) ||
@@ -702,6 +799,9 @@ function removeFiles ( dir_path,extList ) {
 let rc = true;
 
   if (fileExists(dir_path))  {
+
+    flushDirCache ( dir_path );
+
     fs.readdirSync(dir_path).forEach(function(file,index){
       let dlt = false;
       let fl  = file.toLowerCase();
@@ -709,7 +809,7 @@ let rc = true;
         dlt = fl.endsWith(extList[i]);
       if (dlt)  {
         let curPath = path.join ( dir_path,file );
-        let curstat = fileExists ( curPath );
+        let curstat = fileStat  ( curPath );
         if (!curstat)  {
           log.error ( 190,'cannot stat path ' + curPath );
           rc = false;
@@ -725,6 +825,7 @@ let rc = true;
         }
       }
     });
+
   }
 
   return rc;  // false if there were errors
@@ -771,13 +872,16 @@ let html  = '<!DOCTYPE html>\n<html><link rel="stylesheet" type="text/css" ' +
 
 const signal_file_name = 'signal';  // signal file of job termination status
 
+
 function jobSignalExists ( jobDir ) {
   return fileExists ( path.join(jobDir,signal_file_name) );
 }
 
+
 function removeJobSignal ( jobDir ) {
   removeFile ( path.join(jobDir,signal_file_name) );
 }
+
 
 function writeJobSignal ( jobDir,signal_name,signal_message,signal_code )  {
   let line = signal_name;
@@ -785,6 +889,7 @@ function writeJobSignal ( jobDir,signal_name,signal_message,signal_code )  {
     line += ' ' + signal_message;
   writeString ( path.join(jobDir,signal_file_name),line + '\n' + signal_code );
 }
+
 
 function getJobSignalCode ( jobDir )  {
 let code   = 0;
@@ -1022,7 +1127,9 @@ function padDigits ( number,digits ) {
 
 // ==========================================================================
 // export for use in node
+module.exports.configureCache        = configureCache;
 module.exports.fileExists            = fileExists;
+module.exports.fileStat              = fileStat;
 module.exports.isSymbolicLink        = isSymbolicLink;
 module.exports.dirExists             = dirExists;
 module.exports.fileSize              = fileSize;
@@ -1036,6 +1143,7 @@ module.exports.appendString          = appendString;
 module.exports.writeObject           = writeObject;
 module.exports.copyFile              = copyFile;
 module.exports.moveFile              = moveFile;
+module.exports.flushDirCache         = flushDirCache;
 module.exports.moveDir               = moveDir;
 module.exports.moveDirAsync          = moveDirAsync;
 module.exports.copyDirAsync          = copyDirAsync;
